+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/init_decode.h
@@ -25,7 +25,7 @@
  *  Initiation of decoder instance.
  *---------------------------------------------------------------*/
 
-int16_t WebRtcIlbcfix_InitDecode(  /* (o) Number of decoded samples */
+int WebRtcIlbcfix_InitDecode(  /* (o) Number of decoded samples */
     IlbcDecoder *iLBCdec_inst, /* (i/o) Decoder instance */
     int16_t mode,     /* (i) frame size mode */
     int use_enhancer           /* (i) 1 to use enhancer
